test: Move unit test helpers to global test helpers (#4682)
`get_top_level_dir` and `cloud_init_project_dir` have uses in
integration tests. Move them up to a top-level test helper file
accordingly.
