+++ tests/unittests/config/test_schema.py
@@ -43,11 +43,11 @@ from cloudinit.settings import FREQUENCIES
 from cloudinit.sources import DataSourceNotFoundException
 from cloudinit.templater import JinjaSyntaxParsingException
 from cloudinit.util import load_file, write_file
+from tests.helpers import cloud_init_project_dir
 from tests.hypothesis import given
 from tests.hypothesis_jsonschema import from_schema
 from tests.unittests.helpers import (
     CiTestCase,
-    cloud_init_project_dir,
     does_not_raise,
     mock,
     skipUnlessHypothesisJsonSchema,
