+++ tests/unittests/test_render_template.py
@@ -5,7 +5,7 @@ import sys
 import pytest
 
 from cloudinit import subp, templater, util
-from tests.unittests.helpers import cloud_init_project_dir
+from tests.helpers import cloud_init_project_dir
 
 # TODO(Look to align with tools.render-template or cloudinit.distos.OSFAMILIES)
 DISTRO_VARIANTS = [
